docs: consolidate scripts and update documentation - Consolidate migration scripts into mcp-migration.py - Merge enhanced environment verification into verify_environment.py - Merge enhanced site customization into fix_sitecustomize.py - Update documentation with new migration guide and technical details - Rename tag_storage.md to tag-storage.md for consistency
