+++ docs/README.md
@@ -6,10 +6,12 @@ Welcome to the MCP Memory Service documentation. This directory contains compreh
 
 - [Installation Guide](guides/installation.md) - Comprehensive installation instructions for all platforms
 - [Troubleshooting Guide](guides/troubleshooting.md) - Solutions for common issues and debugging procedures
+- [Migration Guide](guides/migration.md) - Instructions for migrating memories between different ChromaDB instances
 
 ## Technical Documentation
 
 - [Tag Storage Procedure](technical/tag-storage.md) - Technical details about tag storage and migration
+- [Memory Migration](technical/memory-migration.md) - Technical details about memory migration process
 
 ## Quick Links
 
@@ -32,11 +34,37 @@ Welcome to the MCP Memory Service documentation. This directory contains compreh
 ## Available Scripts
 
 The `scripts/` directory contains several useful tools:
+
+### Core Scripts
 - `run_memory_server.py` - Direct runner for MCP Memory Service
-- `verify_environment_enhanced.py` - Environment compatibility verification
+- `verify_environment.py` - Enhanced environment compatibility verification
+- `fix_sitecustomize.py` - Fix for recursion issues with enhanced platform support
+- `mcp-migration.py` - Memory migration tool supporting local and remote migrations
+
+### Platform-Specific Scripts
 - `install_windows.py` - Windows-specific installation
-- `fix_sitecustomize.py` - Fix for recursion issues
 - `verify_pytorch_windows.py` - Windows PyTorch verification
+
+## Script Usage Examples
+
+### Environment Verification
+```bash
+python scripts/verify_environment.py
+```
+
+### Memory Migration
+```bash
+# Local to Remote Migration
+python scripts/mcp-migration.py --source-type local --source-config /path/to/local/chroma --target-type remote --target-config '{"host": "remote-host", "port": 8000}'
+
+# Remote to Local Migration
+python scripts/mcp-migration.py --source-type remote --source-config '{"host": "remote-host", "port": 8000}' --target-type local --target-config /path/to/local/chroma
+```
+
+### Site Customization Fix
+```bash
+python scripts/fix_sitecustomize.py
+```
 
 ## Configuration
