+++ scripts/verify_environment.py
@@ -1,25 +1,232 @@
 #!/usr/bin/env python3
-import sys
+"""
+Enhanced environment verification script for MCP Memory Service.
+This script checks the system environment, hardware capabilities,
+and installed dependencies to ensure compatibility.
+"""
 import os
-import pkg_resources
-import importlib
+import sys
 import platform
+import subprocess
 import json
+import importlib
+import pkg_resources
 from pathlib import Path
+import traceback
+import ctypes
 
 class EnvironmentVerifier:
     def __init__(self):
         self.verification_results = []
         self.critical_failures = []
+        self.warnings = []
+        self.system_info = self.detect_system()
+        self.gpu_info = self.detect_gpu()
         self.claude_config = self.load_claude_config()
+
+    def detect_system(self):
+        """Detect system architecture and platform."""
+        system_info = {
+            "os_name": platform.system().lower(),
+            "os_version": platform.version(),
+            "architecture": platform.machine().lower(),
+            "python_version": platform.python_version(),
+            "cpu_count": os.cpu_count() or 1,
+            "memory_gb": self.get_system_memory(),
+            "in_virtual_env": sys.prefix != sys.base_prefix
+        }
+        
+        self.verification_results.append(
+            f"✓ System: {platform.system()} {platform.version()}"
+        )
+        self.verification_results.append(
+            f"✓ Architecture: {system_info['architecture']}"
+        )
+        self.verification_results.append(
+            f"✓ Python: {system_info['python_version']}"
+        )
+        
+        if system_info["in_virtual_env"]:
+            self.verification_results.append(
+                f"✓ Virtual environment: {sys.prefix}"
+            )
+        else:
+            self.warnings.append(
+                "Not running in a virtual environment"
+            )
+        
+        return system_info
+
+    def get_system_memory(self):
+        """Get the total system memory in GB."""
+        try:
+            if self.system_info["os_name"] == "linux":
+                with open('/proc/meminfo', 'r') as f:
+                    for line in f:
+                        if line.startswith('MemTotal:'):
+                            memory_kb = int(line.split()[1])
+                            return round(memory_kb / (1024 * 1024), 2)
+                            
+            elif self.system_info["os_name"] == "darwin":
+                output = subprocess.check_output(['sysctl', '-n', 'hw.memsize']).decode('utf-8').strip()
+                memory_bytes = int(output)
+                return round(memory_bytes / (1024**3), 2)
+                
+            elif self.system_info["os_name"] == "windows":
+                class MEMORYSTATUSEX(ctypes.Structure):
+                    _fields_ = [
+                        ('dwLength', ctypes.c_ulong),
+                        ('dwMemoryLoad', ctypes.c_ulong),
+                        ('ullTotalPhys', ctypes.c_ulonglong),
+                        ('ullAvailPhys', ctypes.c_ulonglong),
+                        ('ullTotalPageFile', ctypes.c_ulonglong),
+                        ('ullAvailPageFile', ctypes.c_ulonglong),
+                        ('ullTotalVirtual', ctypes.c_ulonglong),
+                        ('ullAvailVirtual', ctypes.c_ulonglong),
+                        ('ullAvailExtendedVirtual', ctypes.c_ulonglong),
+                    ]
+                    
+                memoryStatus = MEMORYSTATUSEX()
+                memoryStatus.dwLength = ctypes.sizeof(MEMORYSTATUSEX)
+                ctypes.windll.kernel32.GlobalMemoryStatusEx(ctypes.byref(memoryStatus))
+                return round(memoryStatus.ullTotalPhys / (1024**3), 2)
+                
+        except Exception as e:
+            self.warnings.append(f"Failed to get system memory: {e}")
+            
+        return 4.0  # Conservative default
+
+    def detect_gpu(self):
+        """Detect GPU and acceleration capabilities."""
+        gpu_info = {
+            "has_cuda": False,
+            "cuda_version": None,
+            "has_rocm": False,
+            "rocm_version": None,
+            "has_mps": False,
+            "has_directml": False,
+            "accelerator": "cpu"
+        }
+        
+        # Check for CUDA
+        if self.system_info["os_name"] == "windows":
+            cuda_path = os.environ.get('CUDA_PATH')
+            if cuda_path and os.path.exists(cuda_path):
+                gpu_info["has_cuda"] = True
+                try:
+                    nvcc_output = subprocess.check_output(
+                        [os.path.join(cuda_path, 'bin', 'nvcc'), '--version'],
+                        stderr=subprocess.STDOUT,
+                        universal_newlines=True
+                    )
+                    for line in nvcc_output.split('\n'):
+                        if 'release' in line:
+                            gpu_info["cuda_version"] = line.split('release')[-1].strip().split(',')[0].strip()
+                            break
+                except (subprocess.SubprocessError, FileNotFoundError):
+                    pass
+        elif self.system_info["os_name"] == "linux":
+            cuda_paths = ['/usr/local/cuda', os.environ.get('CUDA_HOME')]
+            for path in cuda_paths:
+                if path and os.path.exists(path):
+                    gpu_info["has_cuda"] = True
+                    try:
+                        nvcc_output = subprocess.check_output(
+                            [os.path.join(path, 'bin', 'nvcc'), '--version'],
+                            stderr=subprocess.STDOUT,
+                            universal_newlines=True
+                        )
+                        for line in nvcc_output.split('\n'):
+                            if 'release' in line:
+                                gpu_info["cuda_version"] = line.split('release')[-1].strip().split(',')[0].strip()
+                                break
+                    except (subprocess.SubprocessError, FileNotFoundError):
+                        pass
+                    break
+        
+        # Check for ROCm
+        if self.system_info["os_name"] == "linux":
+            rocm_paths = ['/opt/rocm', os.environ.get('ROCM_HOME')]
+            for path in rocm_paths:
+                if path and os.path.exists(path):
+                    gpu_info["has_rocm"] = True
+                    try:
+                        with open(os.path.join(path, 'bin', '.rocmversion'), 'r') as f:
+                            gpu_info["rocm_version"] = f.read().strip()
+                    except (FileNotFoundError, IOError):
+                        try:
+                            rocm_output = subprocess.check_output(
+                                ['rocminfo'],
+                                stderr=subprocess.STDOUT,
+                                universal_newlines=True
+                            )
+                            for line in rocm_output.split('\n'):
+                                if 'Version' in line:
+                                    gpu_info["rocm_version"] = line.split(':')[-1].strip()
+                                    break
+                        except (subprocess.SubprocessError, FileNotFoundError):
+                            pass
+                    break
+        
+        # Check for MPS
+        if self.system_info["os_name"] == "darwin" and self.system_info["architecture"] in ("arm64", "aarch64"):
+            try:
+                result = subprocess.run(
+                    ['system_profiler', 'SPDisplaysDataType'],
+                    capture_output=True,
+                    text=True
+                )
+                gpu_info["has_mps"] = 'Metal' in result.stdout
+            except (subprocess.SubprocessError, FileNotFoundError):
+                pass
+        
+        # Check for DirectML
+        if self.system_info["os_name"] == "windows":
+            try:
+                pkg_resources.get_distribution('torch-directml')
+                gpu_info["has_directml"] = True
+            except (ImportError, pkg_resources.DistributionNotFound):
+                try:
+                    ctypes.WinDLL('DirectML.dll')
+                    gpu_info["has_directml"] = True
+                except (ImportError, OSError):
+                    pass
+        
+        # Set accelerator type
+        if gpu_info["has_cuda"]:
+            gpu_info["accelerator"] = "cuda"
+            self.verification_results.append(
+                f"✓ CUDA detected: {gpu_info['cuda_version'] or 'Unknown version'}"
+            )
+        elif gpu_info["has_rocm"]:
+            gpu_info["accelerator"] = "rocm"
+            self.verification_results.append(
+                f"✓ ROCm detected: {gpu_info['rocm_version'] or 'Unknown version'}"
+            )
+        elif gpu_info["has_mps"]:
+            gpu_info["accelerator"] = "mps"
+            self.verification_results.append(
+                "✓ Apple Metal Performance Shaders (MPS) detected"
+            )
+        elif gpu_info["has_directml"]:
+            gpu_info["accelerator"] = "directml"
+            self.verification_results.append(
+                "✓ DirectML detected"
+            )
+        else:
+            self.verification_results.append(
+                "✓ Using CPU-only mode (no GPU acceleration detected)"
+            )
+        
+        return gpu_info
 
     def load_claude_config(self):
         """Load configuration from Claude Desktop config."""
         try:
-            # Check common locations for Claude Desktop config
+            home_dir = Path.home()
             possible_paths = [
-                Path.home() / "Library/Application Support/Claude Desktop/claude_desktop_config.json",
-                Path.home() / ".config/Claude Desktop/claude_desktop_config.json",
+                home_dir / "Library/Application Support/Claude/claude_desktop_config.json",
+                home_dir / ".config/Claude/claude_desktop_config.json",
                 Path(__file__).parent.parent / "claude_config/claude_desktop_config.json"
             ]
 
@@ -32,7 +239,7 @@ class EnvironmentVerifier:
                         )
                         return config
 
-            self.critical_failures.append(
+            self.warnings.append(
                 "Could not find Claude Desktop config file in any standard location"
             )
             return None
@@ -47,7 +254,7 @@ class EnvironmentVerifier:
         """Verify Python interpreter version matches production requirements."""
         try:
             python_version = sys.version.split()[0]
-            required_version = "3.9"  # Production version
+            required_version = "3.10"  # Updated to match current requirements
             
             if not python_version.startswith(required_version):
                 self.critical_failures.append(
@@ -110,7 +317,6 @@ class EnvironmentVerifier:
             return
 
         try:
-            # Extract paths from config
             chroma_path = self.claude_config.get('mcp-memory', {}).get('chroma_db')
             backup_path = self.claude_config.get('mcp-memory', {}).get('backup_path')
 
@@ -173,7 +379,7 @@ class EnvironmentVerifier:
         self.verify_python_version()
         self.verify_virtual_environment()
         self.verify_critical_packages()
-        self.verify_claude_paths()  # This now sets environment variables from config
+        self.verify_claude_paths()
         self.verify_import_functionality()
         self.verify_paths()
 
@@ -186,6 +392,11 @@ class EnvironmentVerifier:
             for result in self.verification_results:
                 print(f"  {result}")
         
+        if self.warnings:
+            print("\nWarnings:")
+            for warning in self.warnings:
+                print(f"  ⚠️  {warning}")
+        
         if self.critical_failures:
             print("\nCritical Failures:")
             for failure in self.critical_failures:
@@ -193,13 +404,14 @@ class EnvironmentVerifier:
         
         print("\nSummary:")
         print(f"  Passed: {len(self.verification_results)}")
+        print(f"  Warnings: {len(self.warnings)}")
         print(f"  Failed: {len(self.critical_failures)}")
         
         if self.critical_failures:
             print("\nTo fix these issues:")
             print("1. Create a new virtual environment:")
-            print("   conda create -n migration-env python=3.9")
-            print("   conda activate migration-env")
+            print("   conda create -n mcp-env python=3.10")
+            print("   conda activate mcp-env")
             print("\n2. Install requirements:")
             print("   pip install -r requirements.txt")
             print("\n3. Ensure Claude Desktop config is properly set up with required paths")
@@ -212,10 +424,10 @@ def main():
     environment_ok = verifier.print_results()
     
     if not environment_ok:
-        print("\n⚠️  Environment verification failed! Migration cannot proceed.")
+        print("\n⚠️  Environment verification failed! Please fix the issues above.")
         sys.exit(1)
     else:
-        print("\n✓ Environment verification passed! Safe to proceed with migration.")
+        print("\n✓ Environment verification passed! Safe to proceed.")
         sys.exit(0)
 
 if __name__ == "__main__":
